PLANET-8233 Retire old Timeline block
This switches to the new design by removing the feature flag and unnecessary code
