+++ assets/src/blocks/Timeline/TimelineBlock.js
@@ -1,14 +1,9 @@
 import {renderToString} from 'react-dom/server';
 import {TimelineEditor} from './TimelineEditorScript';
-import {NewTimelineEditor} from './NewTimelineEditorScript';
 import {TimelineFrontend} from './TimelineFrontend';
-import {NewTimelineFrontend} from './NewTimelineFrontend';
 import {frontendRendered} from '../../functions/frontendRendered';
-import {getUniqueId} from '../../functions/getUniqueId';
 
 const BLOCK_NAME = 'planet4-blocks/timeline';
-
-const newTimelineEnabled = Boolean(window.p4_vars.features.new_timeline_block);
 
 export const registerTimelineBlock = () => {
   const {registerBlockType} = wp.blocks;
@@ -16,7 +11,7 @@ export const registerTimelineBlock = () => {
   const {RawHTML} = wp.element;
   const {useBlockProps} = wp.blockEditor;
 
-  const blockAttributes = {
+  const attributes = {
     timeline_title: {
       type: 'string',
       default: '',
@@ -41,6 +36,10 @@ export const registerTimelineBlock = () => {
       type: 'boolean',
       default: false,
     },
+    timeline_id: {
+      type: 'string',
+      default: '',
+    },
   };
 
   registerBlockType(BLOCK_NAME, {
@@ -51,13 +50,10 @@ export const registerTimelineBlock = () => {
     supports: {
       html: false, // Disable "Edit as HTMl" block option.
     },
-    attributes: newTimelineEnabled ? {
-      ...blockAttributes,
-      timeline_id: {type: 'string', default: ''},
-    } : blockAttributes,
+    attributes,
     edit: props => (
       <div {...useBlockProps()}>
-        {newTimelineEnabled ? <NewTimelineEditor {...props} /> : <TimelineEditor {...props} />}
+        <TimelineEditor {...props} />
       </div>
     ),
     save: props => {
@@ -66,42 +62,18 @@ export const registerTimelineBlock = () => {
           data-hydrate={BLOCK_NAME}
           data-attributes={JSON.stringify(props.attributes)}
         >
-          {newTimelineEnabled ? <NewTimelineFrontend {...props} /> : <TimelineFrontend {...props} />}
+          <TimelineFrontend {...props} />
         </div>
       );
       return <RawHTML>{markup}</RawHTML>;
     },
     deprecated: [
-      newTimelineEnabled ? {
-        attributes: blockAttributes,
-        edit: TimelineEditor,
-        save: props => {
-          const markup = renderToString(
-            <div
-              data-hydrate={BLOCK_NAME}
-              data-attributes={JSON.stringify(props.attributes)}
-            >
-              <TimelineFrontend {...props} />
-            </div>
-          );
-          return <RawHTML>{markup}</RawHTML>;
-        },
-        isEligible(attributes) {
-          return !attributes.timeline_id;
-        },
-        migrate(attributes) {
-          return {
-            ...attributes,
-            timeline_id: getUniqueId('tl'),
-          };
-        },
-      } : {},
       {
-        attributes: blockAttributes,
+        attributes,
         save: frontendRendered(BLOCK_NAME),
       },
       {
-        attributes: blockAttributes,
+        attributes,
         save() {
           return null;
         },
