+++ assets/src/blocks/Timeline/TimelineEditorScript.js
@@ -1,159 +1,123 @@
 import {URLInput} from '../../block-editor/URLInput/URLInput';
-import {useScript} from '../../hooks/useScript/useScript';
-import {useStyleSheet} from '../../hooks/useStylesheet/useStylesheet';
-import {Timeline} from './Timeline';
-import {languages} from './TimelineLanguages';
+import {TimelineFrontend} from './TimelineFrontend';
 import {URLDescriptionHelp} from './URLDescriptionHelp';
-import {isLodash} from '../../functions/isLodash';
+import {getUniqueId} from '../../functions/getUniqueId';
 
 const {InspectorControls, RichText} = wp.blockEditor;
-const {PanelBody, SelectControl, CheckboxControl, Tooltip} = wp.components;
+const {PanelBody, Tooltip} = wp.components;
+import {select} from '@wordpress/data';
 const {debounce} = wp.compose;
-const {useCallback, useState} = wp.element;
+const {useCallback, useState, useEffect} = wp.element;
 const {__} = wp.i18n;
 
-const TIMELINE_JS_VERSION = '3.8.12';
+/**
+ * Get all blocks and inner blocks from a page.
+ *
+ * @param {Array} blocks - an array of blocks.
+ *
+ * @return {Array} All blocks from the page.
+ */
+const flattenBlocks = blocks => blocks.flatMap(block => [block, ...flattenBlocks(block.innerBlocks || [])]);
 
-const positions = [
-  {label: 'Bottom', value: 'bottom'},
-  {label: 'Top', value: 'top'},
-];
-
-const loadAssets = () => {
-  const revertLodash = function() {
-    // Address conflicts between the underscore and lodash libraries.
-    if (isLodash()) {
-      // eslint-disable-next-line no-undef
-      _.noConflict();
-    }
-  };
-  // eslint-disable-next-line no-unused-vars
-  const [scriptLoaded, scriptError] = useScript(
-    `https://cdn.knightlab.com/libs/timeline3/${TIMELINE_JS_VERSION}/js/timeline-min.js`,
-    revertLodash
+/**
+ * Check if the given timeline id already exists on this page.
+ * This can happen if we duplicate a block for example.
+ *
+ * @param {string} timelineId - the Timeline block id.
+ *
+ * @return {boolean} Whether the timeline id is already given or not.
+ */
+const isTimelineIdReserved = timelineId => {
+  const allBlocks = flattenBlocks(select('core/editor').getBlocks());
+  return allBlocks.some(
+    block => block.name === 'planet4-blocks/timeline' && block.attributes.timeline_id === timelineId
   );
-
-  // eslint-disable-next-line no-unused-vars
-  const [stylesLoaded, stylesError] = useStyleSheet(
-    `https://cdn.knightlab.com/libs/timeline3/${TIMELINE_JS_VERSION}/css/timeline.css`
-  );
-
-  return [scriptLoaded, stylesLoaded];
 };
 
 const renderEdit = (
-  {language, timenav_position, start_at_end},
-  toAttribute,
   sheetURL,
   setSheetURL,
   debounceSheetURLUpdate
-) => {
-  return (
-    <InspectorControls>
-      <PanelBody title={__('Settings', 'planet4-master-theme-backend')}>
-        <URLInput
-          label={__('Google Sheets URL', 'planet4-master-theme-backend')}
-          placeholder={__('Enter URL', 'planet4-master-theme-backend')}
-          value={sheetURL}
-          onChange={value => {
-            setSheetURL(value);
-            debounceSheetURLUpdate(value);
-          }}
-        />
-
-        <URLDescriptionHelp />
-
-        <SelectControl
-          __nextHasNoMarginBottom
-          __next40pxDefaultSize
-          label={__('Language', 'planet4-master-theme-backend')}
-          value={language}
-          options={languages}
-          onChange={toAttribute('language')}
-        />
-
-        <SelectControl
-          __nextHasNoMarginBottom
-          __next40pxDefaultSize
-          label={__('Timeline navigation position', 'planet4-master-theme-backend')}
-          value={timenav_position}
-          options={positions}
-          onChange={toAttribute('timenav_position')}
-        />
-
-        <CheckboxControl
-          __nextHasNoMarginBottom
-          label={__('Start at end', 'planet4-master-theme-backend')}
-          help={__('Begin at the end of the timeline', 'planet4-master-theme-backend')}
-          value={start_at_end}
-          checked={start_at_end}
-          onChange={toAttribute('start_at_end')}
-        />
-
-      </PanelBody>
-      <PanelBody title={__('Learn more about this block', 'planet4-master-theme-backend')} initialOpen={false}>
-        <p className="components-base-control__help">
-          <a target="_blank" href="https://planet4.greenpeace.org/content/blocks/timeline/" rel="noreferrer">
-            P4 Handbook Timeline
-          </a>
-          {' '} &#8987;
-        </p>
-      </PanelBody>
-    </InspectorControls>
-  );
-};
-
-const renderView = (attributes, toAttribute, scriptLoaded, stylesLoaded) => {
-  return (
-    <section className={`block timeline-block ${attributes.className ?? ''}`}>
-      <Tooltip text={__('Edit text', 'planet4-master-theme-backend')}>
-        <header>
-          <RichText
-            tagName="h2"
-            className="page-section-header"
-            placeholder={__('Enter title', 'planet4-master-theme-backend')}
-            value={attributes.timeline_title}
-            onChange={toAttribute('timeline_title')}
-            withoutInteractiveFormatting
-            maxLength={40}
-            allowedFormats={[]}
-          />
-        </header>
-      </Tooltip>
-      <RichText
-        tagName="p"
-        className="page-section-description"
-        placeholder={__('Enter description', 'planet4-master-theme-backend')}
-        value={attributes.description}
-        onChange={toAttribute('description')}
-        withoutInteractiveFormatting
-        maxLength={200}
-        allowedFormats={['core/bold', 'core/italic']}
+) => (
+  <InspectorControls>
+    <PanelBody title={__('Settings', 'planet4-master-theme-backend')}>
+      <URLInput
+        label={__('Google Sheets URL', 'planet4-master-theme-backend')}
+        placeholder={__('Enter URL', 'planet4-master-theme-backend')}
+        value={sheetURL}
+        onChange={value => {
+          setSheetURL(value);
+          debounceSheetURLUpdate(value);
+        }}
       />
-      {!attributes.google_sheets_url &&
-        <div className="block-edit-mode-warning components-notice is-warning">
-          { __('Please include a Sheet URL.', 'planet4-master-theme-backend') }
-        </div>
-      }
-      {attributes.google_sheets_url && scriptLoaded && stylesLoaded &&
-        <Timeline {...attributes} />
-      }
-    </section>
-  );
-};
+
+      <URLDescriptionHelp />
+
+    </PanelBody>
+    <PanelBody title={__('Learn more about this block', 'planet4-master-theme-backend')} initialOpen={false}>
+      <p className="components-base-control__help">
+        <a target="_blank" href="https://planet4.greenpeace.org/content/blocks/timeline/" rel="noreferrer">
+          P4 Handbook Timeline
+        </a>
+        {' '} &#8987;
+      </p>
+    </PanelBody>
+  </InspectorControls>
+);
+
+const renderView = (attributes, toAttribute) => (
+  <section className={`block timeline-block ${attributes.className ?? ''}`}>
+    <Tooltip text={__('Edit text', 'planet4-master-theme-backend')}>
+      <header>
+        <RichText
+          tagName="h2"
+          className="page-section-header"
+          placeholder={__('Enter title', 'planet4-master-theme-backend')}
+          value={attributes.timeline_title}
+          onChange={toAttribute('timeline_title')}
+          withoutInteractiveFormatting
+          maxLength={40}
+          allowedFormats={[]}
+        />
+      </header>
+    </Tooltip>
+    <RichText
+      tagName="p"
+      className="page-section-description"
+      placeholder={__('Enter description', 'planet4-master-theme-backend')}
+      value={attributes.description}
+      onChange={toAttribute('description')}
+      withoutInteractiveFormatting
+      maxLength={200}
+      allowedFormats={['core/bold', 'core/italic']}
+    />
+    {!attributes.google_sheets_url ?
+      <div className="block-edit-mode-warning components-notice is-warning">
+        { __('Please include a Sheet URL.', 'planet4-master-theme-backend') }
+      </div> :
+      <TimelineFrontend attributes={{isEditing: true, ...attributes}} />
+    }
+  </section>
+);
 
 export const TimelineEditor = ({isSelected, attributes, setAttributes}) => {
   const toAttribute = attributeName => value => setAttributes({[attributeName]: value});
-  const [scriptLoaded, stylesLoaded] = loadAssets();
   // Using a state to prevent the input losing the cursor position, a React issue reported multiple times
   const [sheetURL, setSheetURL] = useState(attributes.google_sheets_url);
   const debounceSheetURLUpdate = useCallback(debounce(toAttribute('google_sheets_url'), 300), []);
 
+  // Set a timeline id in the attributes, if it doesn't exist yet.
+  useEffect(() => {
+    if (!attributes.timeline_id || isTimelineIdReserved(attributes.timeline_id)) {
+      setAttributes({timeline_id: getUniqueId('tl')});
+    }
+  }, []);
+
   return (
     <>
-      {renderView(attributes, toAttribute, scriptLoaded, stylesLoaded)}
+      {renderView(attributes, toAttribute)}
       {isSelected &&
-        renderEdit(attributes, toAttribute, sheetURL, setSheetURL, debounceSheetURLUpdate)
+        renderEdit(sheetURL, setSheetURL, debounceSheetURLUpdate)
       }
     </>
   );
