--- assets/src/blocks/Timeline/TimelineLanguages.js
@@ -1,64 +1,0 @@
-export const languages = [
-  {label: 'Afrikaans', value: 'af'},
-  {label: 'Arabic', value: 'ar'},
-  {label: 'Armenian', value: 'hy'},
-  {label: 'Basque', value: 'eu'},
-  {label: 'Belarusian', value: 'be'},
-  {label: 'Bulgarian', value: 'bg'},
-  {label: 'Catalan', value: 'ca'},
-  {label: 'Chinese', value: 'zh-cn'},
-  {label: 'Croatian / Hrvatski', value: 'hr'},
-  {label: 'Czech', value: 'cz'},
-  {label: 'Danish', value: 'da'},
-  {label: 'Dutch', value: 'nl'},
-  {label: 'English', value: 'en'},
-  {label: 'English (24-hour time)', value: 'en-24hr'},
-  {label: 'Esperanto', value: 'eo'},
-  {label: 'Estonian', value: 'et'},
-  {label: 'Faroese', value: 'fo'},
-  {label: 'Farsi', value: 'fa'},
-  {label: 'Finnish', value: 'fi'},
-  {label: 'French', value: 'fr'},
-  {label: 'Frisian', value: 'fy'},
-  {label: 'Galician', value: 'gl'},
-  {label: 'Georgian', value: 'ka'},
-  {label: 'German / Deutsch', value: 'de'},
-  {label: 'Greek', value: 'el'},
-  {label: 'Hebrew', value: 'he'},
-  {label: 'Hindi', value: 'hi'},
-  {label: 'Hungarian', value: 'hu'},
-  {label: 'Icelandic', value: 'is'},
-  {label: 'Indonesian', value: 'id'},
-  {label: 'Irish', value: 'ga'},
-  {label: 'Italian', value: 'it'},
-  {label: 'Japanese', value: 'ja'},
-  {label: 'Korean', value: 'ko'},
-  {label: 'Latvian', value: 'lv'},
-  {label: 'Lithuanian', value: 'lt'},
-  {label: 'Luxembourgish', value: 'lb'},
-  {label: 'Malay', value: 'ms'},
-  {label: 'Myanmar', value: 'my'},
-  {label: 'Nepali', value: 'ne'},
-  {label: 'Norwegian', value: 'no'},
-  {label: 'Polish', value: 'pl'},
-  {label: 'Portuguese', value: 'pt'},
-  {label: 'Portuguese (Brazilian)', value: 'pt-br'},
-  {label: 'Romanian', value: 'ro'},
-  {label: 'Romansh', value: 'rm'},
-  {label: 'Russian', value: 'ru'},
-  {label: 'Serbian - Cyrillic', value: 'sr-cy'},
-  {label: 'Serbian - Latin', value: 'sr'},
-  {label: 'Sinhalese', value: 'si'},
-  {label: 'Slovak', value: 'sk'},
-  {label: 'Slovenian', value: 'sl'},
-  {label: 'Spanish', value: 'es'},
-  {label: 'Swedish', value: 'sv'},
-  {label: 'Tagalog', value: 'tl'},
-  {label: 'Tamil', value: 'ta'},
-  {label: 'Taiwanese', value: 'zh-tw'},
-  {label: 'Telugu', value: 'te'},
-  {label: 'Thai', value: 'th'},
-  {label: 'Turkish', value: 'tr'},
-  {label: 'Ukrainian', value: 'uk'},
-  {label: 'Urdu', value: 'ur'},
-];
