+++ assets/src/scss/blocks/Timeline/TimelineEditorStyle.scss
@@ -1,13 +1,4 @@
 div[data-type="planet4-blocks/timeline"] {
-  height: 890px;
-
-  .page-section-header,
-  .page-section-description {
-    text-align: center;
-  }
-}
-
-.new-timeline-block {
   max-height: 100%;
   overflow: hidden;
