+++ src/Blocks/Timeline.php
@@ -9,8 +9,6 @@
 
 namespace P4\MasterTheme\Blocks;
 
-use P4\MasterTheme\Features\Dev\NewTimelineBlock;
-
 /**
  * Class Timeline
  *
@@ -25,26 +23,6 @@ class Timeline extends BaseBlock
     public const BLOCK_NAME = 'timeline';
 
     /**
-     * @const string TIMELINE_JS_VERSION.
-     */
-    public const TIMELINE_JS_VERSION = '3.8.12';
-
-    /**
-     * @const string TIMELINE_LIB.
-     */
-    public const TIMELINE_LIB = 'https://cdn.knightlab.com/libs/timeline3/';
-
-    /**
-     * @const string TIMELINE_JS_SCRIPT.
-     */
-    public const TIMELINE_JS_SCRIPT = 'timeline-js';
-
-    /**
-     * @const string TIMELINE_CSS_SCRIPT.
-     */
-    public const TIMELINE_CSS_SCRIPT = 'timeline-css';
-
-    /**
      * Timeline constructor.
      */
     public function __construct()
@@ -57,40 +35,6 @@ class Timeline extends BaseBlock
      */
     public function register_timeline_block(): void
     {
-        $attributes = [
-            'timeline_title' => [
-                'type' => 'string',
-                'default' => '',
-            ],
-            'description' => [
-                'type' => 'string',
-                'default' => '',
-            ],
-            'google_sheets_url' => [
-                'type' => 'string',
-                'default' => '',
-            ],
-            'language' => [
-                'type' => 'string',
-                'default' => 'en',
-            ],
-            'timenav_position' => [
-                'type' => 'string',
-                'default' => 'bottom',
-            ],
-            'start_at_end' => [
-                'type' => 'boolean',
-                'default' => false,
-            ],
-        ];
-
-        if (NewTimelineBlock::is_active()) {
-            $attributes['timeline_id'] = [
-                'type' => 'string',
-                'default' => '',
-            ];
-        }
-
         // - Register the block for the editor
         // in the PHP side.
         register_block_type(
@@ -100,65 +44,43 @@ class Timeline extends BaseBlock
                 'editor_script' => 'planet4-blocks',
                 // todo: Remove when all content is migrated.
                 'render_callback' => [ self::class, 'hydrate_frontend' ],
-                'attributes' => $attributes,
+                'attributes' => [
+                    'timeline_title' => [
+                        'type' => 'string',
+                        'default' => '',
+                    ],
+                    'description' => [
+                        'type' => 'string',
+                        'default' => '',
+                    ],
+                    'google_sheets_url' => [
+                        'type' => 'string',
+                        'default' => '',
+                    ],
+                    'language' => [
+                        'type' => 'string',
+                        'default' => 'en',
+                    ],
+                    'timenav_position' => [
+                        'type' => 'string',
+                        'default' => 'bottom',
+                    ],
+                    'start_at_end' => [
+                        'type' => 'boolean',
+                        'default' => false,
+                    ],
+                    'timeline_id' => [
+                        'type' => 'string',
+                        'default' => '',
+                    ],
+                ],
             ]
         );
-
-        if (!NewTimelineBlock::is_active()) {
-            wp_register_script(
-                self::TIMELINE_JS_SCRIPT,
-                self::TIMELINE_LIB . self::TIMELINE_JS_VERSION . '/js/timeline-min.js',
-                [],
-                self::TIMELINE_JS_VERSION,
-                true
-            );
-
-            wp_register_style(
-                self::TIMELINE_CSS_SCRIPT,
-                self::TIMELINE_LIB . self::TIMELINE_JS_VERSION . '/css/timeline.css',
-                [],
-                self::TIMELINE_JS_VERSION
-            );
-        }
 
         if (is_admin()) {
             add_action('enqueue_block_assets', [ self::class, 'enqueue_editor_assets' ]);
         }
         add_action('wp_enqueue_scripts', [ self::class, 'enqueue_frontend_assets' ]);
-    }
-
-    /**
-     * Frontend script
-     */
-    public static function enqueue_frontend_script(): void
-    {
-        $deps = ['planet4-blocks-theme-script'];
-
-        if (wp_script_is(self::TIMELINE_JS_SCRIPT, 'registered')) {
-            wp_enqueue_script(self::TIMELINE_JS_SCRIPT);
-            $deps[] = self::TIMELINE_JS_SCRIPT;
-        }
-
-        wp_enqueue_script(
-            static::get_full_block_name() . '-script',
-            static::get_url_path() . 'Script.js',
-            $deps,
-            \P4\MasterTheme\Loader::theme_file_ver(static::get_rel_path() . 'Script.js'),
-            true
-        );
-    }
-
-    /**
-     * Frontend style
-     */
-    public static function enqueue_frontend_style(): void
-    {
-        wp_enqueue_style(
-            static::get_full_block_name() . '-style',
-            static::get_url_path() . 'Style.min.css',
-            [self::TIMELINE_CSS_SCRIPT],
-            \P4\MasterTheme\Loader::theme_file_ver(static::get_rel_path() . 'Style.min.css'),
-        );
     }
 
     /**
